# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.0",
    "title": "DataPorts API with swagger",
    "description": "This is a simple REST API made with Node.JS and documented with Swagger",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "José A. Clemente",
      "email": "jclemente@prodevelop.es"
    }
  },
  "basePath": "/dac",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/images/{id}/template": {
      "get": {
        "tags": [
          "Images"
        ],
        "operationId": "getTemplate",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "summary": "Returns a template for creating an agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the image that we want to retrieve its template",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
    "/pythonTemplate": {
      "get": {
        "tags": [
          "PythonTemplate"
        ],
        "summary": "Retrieves a list of registered python templates",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "succesful operation"
          }
        }
      }
    },
    "/pythonTemplate/getZIPTemplate": {
      "post": {
        "tags": [
          "PythonTemplate"
        ],
        "summary": "Retrieves a list of registered python templates",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "produces": [
          "application/zip"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Information with the zip contents",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "constants",
                "script",
                "dockerFile"
              ],
              "properties": {
                "constants": {
                  "type": "string"
                },
                "script": {
                  "type": "string"
                },
                "dockerFile": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "bearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token with the `Bearer` prefix, E.g., \"Bearer easkkassjejm.dadsaeea.ssdaee\""
    }
  },
  "definitions": {
...
    "GenericMessage": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "GenericAnswer": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}